Fix PCR selection in quote generation.
Quote_In struct was not being zeroed, resulting in an indeterminate
starting state. It's fields were then individually set to ensure correct
configuration before obtaining the quote. However, TPM2_SetupPCRSel that
is used to configure PCR selection only sets up the selection for the
relevant index, and does not touch the other indices. This resulted in
random PRCs being "selected" in addition to the one explicitly set.

Zero Quote_In before populating it to avoid this issue.

Signed-off-by: setrofim <setrofim@gmail.com>